# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cabo Shuttle Ride API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Zone {
  id: number;
  name: string;
  slug: string;
  description: string;
  coverImage: string;
  highlights?: string[];
}

export interface Vehicle {
  id: number;
  name: string;
  slug: string;
  capacity: number;
  description: string;
  amenities: string[];
  imageUrl: string;
  priceModifier?: number;
}

export interface PricingEntry {
  id: number;
  originZoneId: number;
  destinationZoneId: number;
  vehicleId: number;
  priceUsd: number;
  durationMinutes: number;
  originZone?: Zone;
  destinationZone?: Zone;
  vehicle?: Vehicle;
}

export interface ReservationInput {
  passengerName: string;
  passengerEmail: string;
  passengerPhone: string;
  originZoneId: number;
  destinationZoneId: number;
  vehicleId: number;
  pickupAddress: string;
  dropoffAddress: string;
  pickupDatetime: string;
  passengerCount: number;
  /** @nullable */
  flightNumber?: string | null;
  /** @nullable */
  specialRequests?: string | null;
}

export interface Reservation {
  id: number;
  confirmationCode: string;
  passengerName: string;
  passengerEmail: string;
  passengerPhone: string;
  originZoneId: number;
  destinationZoneId: number;
  vehicleId: number;
  pickupAddress: string;
  dropoffAddress: string;
  pickupDatetime: string;
  passengerCount: number;
  /** @nullable */
  flightNumber?: string | null;
  /** @nullable */
  specialRequests?: string | null;
  totalPriceUsd: number;
  status: string;
  createdAt: string;
  originZone?: Zone;
  destinationZone?: Zone;
  vehicle?: Vehicle;
}

export interface ReservationStats {
  total: number;
  confirmed: number;
  cancelled: number;
  pending: number;
  totalRevenue: number;
  todayBookings: number;
  /** @nullable */
  popularRoute: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type ListPricingParams = {
/**
 * @nullable
 */
originZoneId?: number | null;
/**
 * @nullable
 */
destinationZoneId?: number | null;
/**
 * @nullable
 */
vehicleId?: number | null;
};

export type ListReservationsParams = {
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
limit?: number | null;
};