# lib/api-zod/src/generated/types/pricingEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cabo Shuttle Ride API
 * OpenAPI spec version: 0.1.0
 */
import type { Vehicle } from './vehicle';
import type { Zone } from './zone';

export interface PricingEntry {
  id: number;
  originZoneId: number;
  destinationZoneId: number;
  vehicleId: number;
  priceUsd: number;
  durationMinutes: number;
  originZone?: Zone;
  destinationZone?: Zone;
  vehicle?: Vehicle;
}