# lib/api-zod/src/generated/types/reservation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cabo Shuttle Ride API
 * OpenAPI spec version: 0.1.0
 */
import type { Vehicle } from './vehicle';
import type { Zone } from './zone';

export interface Reservation {
  id: number;
  confirmationCode: string;
  passengerName: string;
  passengerEmail: string;
  passengerPhone: string;
  originZoneId: number;
  destinationZoneId: number;
  vehicleId: number;
  pickupAddress: string;
  dropoffAddress: string;
  pickupDatetime: string;
  passengerCount: number;
  /** @nullable */
  flightNumber?: string | null;
  /** @nullable */
  specialRequests?: string | null;
  totalPriceUsd: number;
  status: string;
  createdAt: string;
  originZone?: Zone;
  destinationZone?: Zone;
  vehicle?: Vehicle;
}