# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cabo Shuttle Ride API
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  passengerName: string;
  passengerEmail: string;
  passengerPhone: string;
  originZoneId: number;
  destinationZoneId: number;
  vehicleId: number;
  pickupAddress: string;
  dropoffAddress: string;
  pickupDatetime: string;
  passengerCount: number;
  /** @nullable */
  flightNumber?: string | null;
  /** @nullable */
  specialRequests?: string | null;
}