# lib/api-zod/src/generated/types/reservationStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cabo Shuttle Ride API
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationStats {
  total: number;
  confirmed: number;
  cancelled: number;
  pending: number;
  totalRevenue: number;
  todayBookings: number;
  /** @nullable */
  popularRoute: string | null;
}