# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cabo Shuttle Ride API
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: number;
  name: string;
  slug: string;
  capacity: number;
  description: string;
  amenities: string[];
  imageUrl: string;
  priceModifier?: number;
}