# lib/api-zod/src/generated/types/zone.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cabo Shuttle Ride API
 * OpenAPI spec version: 0.1.0
 */

export interface Zone {
  id: number;
  name: string;
  slug: string;
  description: string;
  coverImage: string;
  highlights?: string[];
}